// src/core/http/generated/models/paymentGatewayProvider.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * Petvex Admin API
 * OpenAPI spec version: 1.0.0
 */

export type PaymentGatewayProvider = typeof PaymentGatewayProvider[keyof typeof PaymentGatewayProvider];


export const PaymentGatewayProvider = {
  mercado_pago: 'mercado_pago',
  pagseguro: 'pagseguro',
  iugu: 'iugu',
  stripe: 'stripe',
  asaas: 'asaas',
} as const;